requirements.txt
fastapi
uvicorn
python-multipart
click
pillow
jinja2
opencv-python-headless
numpy